Remove Instagram section and mention of Instagram in Redirection section
Both of the listed front-ends for Instagram are dead, the fork for Barinsta has been dead for months (https://blog.avalos.me/2021/08/14/barinsta/), and Bibliogram just closed a few weeks ago (https://bibliogram.art/). All remaining public instances that are still up for Bibliogram lead to a 504 Gateway Time-out when trying to search anything.

All apps and extensions in the Redirection section use Bibliogram, so their mention of Instagram should be removed.